backend/requirements.txt
asgiref==3.11.1
Django==4.2.9
django-allauth==0.63.3
django-cors-headers==4.3.1
django-otp==1.7.0
django-ratelimit==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
packaging==26.0
pillow==12.1.1
PyJWT==2.12.1
pypng==0.20220715.0
pytz==2026.1.post1
qrcode==7.4.2
sqlparse==0.5.5
typing_extensions==4.15.0
